requirements.txt
streamlit>=1.28.0
groq>=0.4.1
pandas>=1.5.0
SpeechRecognition>=3.10.0
python-dotenv>=1.0.0
pyaudio>=0.2.11
